requirements.txt
requests==2.31.0
peewee==3.15.0
python-dotenv==0.20.0
pytz==2023.3
dateparser==1.1.8
jsonschema==4.21.1
Pillow==10.3.0
discord.py==2.4.0